+++ cocofest/examples/identification/hmed2018_model_identification.py
@@ -16,6 +16,7 @@ from cocofest import (
     DingModelPulseIntensityFrequencyForceParameterIdentification,
     IvpFes,
 )
+from cocofest.identification.identification_method import full_data_extraction
 
 
 # --- Setting simulation parameters --- #
@@ -54,7 +55,7 @@ result, time = sol_from_initial_guess.integrate(
 )
 
 # Adding noise to the force
-noise = np.random.normal(0, 5, len(result["F"][0]))
+noise = np.random.normal(0, 1, len(result["F"][0]))
 force = result["F"][0] + noise
 
 stim = [final_time / n_stim * i for i in range(n_stim)]
@@ -76,7 +77,7 @@ ocp = DingModelPulseIntensityFrequencyForceParameterIdentification(
     model=model,
     data_path=[pickle_file_name],
     identification_method="full",
-    identification_with_average_method_initial_guess=False,
+    double_step_identification=False,
     key_parameter_to_identify=["a_rest", "km_rest", "tau1_rest", "tau2", "ar", "bs", "Is", "cr"],
     additional_key_settings={},
     n_shooting=n_shooting,
@@ -138,7 +139,7 @@ identified_force = identified_result["F"][0]
     pickle_stim_apparition_time,
     pickle_muscle_data,
     pickle_discontinuity_phase_list,
-) = DingModelPulseIntensityFrequencyForceParameterIdentification.full_data_extraction([pickle_file_name])
+) = full_data_extraction([pickle_file_name])
 
 result_dict = {
     "a_rest": [identified_model.a_rest, DingModelIntensityFrequency().a_rest],
